jetson-edge/pyproject.toml
[project]
name = "detech-jetson-edge"
version = "0.1.0"
description = "DETECH Jetson Edge - YOLO11 object detection and MediaPipe gesture recognition"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "paho-mqtt>=1.6.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"
